pyproject.toml
[project]
name = "shopbot-admin"
version = "0.1.0"
description = ""
authors = [
    {name = "aemareev",email = "aemareev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
